// services/playgame/engine/manifest/location-sets/core-v1/locations.generated.ts
export const CORE_V1_LOCATION_SET = CoreV1LocationSet as unknown as LocationSetMetadata;

export const CORE_V1_LOCATION_MODULES: readonly { folder: string; location: AuthoredLocationDef }[] = [
  { folder: 'ammo-club', location: AmmoClubLocation as unknown as AuthoredLocationDef },
  { folder: 'backdoor', location: BackdoorLocation as unknown as AuthoredLocationDef },
  { folder: 'black-clinic', location: BlackClinicLocation as unknown as AuthoredLocationDef },
  { folder: 'black-halo', location: BlackHaloLocation as unknown as AuthoredLocationDef },
  { folder: 'charging-station', location: ChargingStationLocation as unknown as AuthoredLocationDef },
  { folder: 'chip-tune', location: ChipTuneLocation as unknown as AuthoredLocationDef },
  { folder: 'chrome-beach', location: ChromeBeachLocation as unknown as AuthoredLocationDef },
  { folder: 'chrome-depot', location: ChromeDepotLocation as unknown as AuthoredLocationDef },
  { folder: 'civil-court', location: CivilCourtLocation as unknown as AuthoredLocationDef },
  { folder: 'courthouse', location: CourthouseLocation as unknown as AuthoredLocationDef },
  { folder: 'cryo-storage', location: CryoStorageLocation as unknown as AuthoredLocationDef },
  { folder: 'cryobank', location: CryobankLocation as unknown as AuthoredLocationDef },
  { folder: 'dark-alley', location: DarkAlleyLocation as unknown as AuthoredLocationDef },
  { folder: 'data-chapel', location: DataChapelLocation as unknown as AuthoredLocationDef },
  { folder: 'debt-alley', location: DebtAlleyLocation as unknown as AuthoredLocationDef },
  { folder: 'drone-hive', location: DroneHiveLocation as unknown as AuthoredLocationDef },
  { folder: 'federal-courthouse', location: FederalCourthouseLocation as unknown as AuthoredLocationDef },
  { folder: 'gun-store', location: GunStoreLocation as unknown as AuthoredLocationDef },
  { folder: 'helixdyne-lab', location: HelixdyneLabLocation as unknown as AuthoredLocationDef },
  { folder: 'kurotek-atrium', location: KurotekAtriumLocation as unknown as AuthoredLocationDef },
  { folder: 'meridian-tower', location: MeridianTowerLocation as unknown as AuthoredLocationDef },
  { folder: 'neon-reliquary', location: NeonReliquaryLocation as unknown as AuthoredLocationDef },
  { folder: 'netrunner-shrine', location: NetrunnerShrineLocation as unknown as AuthoredLocationDef },
  { folder: 'organ-bank', location: OrganBankLocation as unknown as AuthoredLocationDef },
  { folder: 'overclock-room', location: OverclockRoomLocation as unknown as AuthoredLocationDef },
  { folder: 'pawn-shop', location: PawnShopLocation as unknown as AuthoredLocationDef },
  { folder: 'power-sink', location: PowerSinkLocation as unknown as AuthoredLocationDef },
  { folder: 'public-pool', location: PublicPoolLocation as unknown as AuthoredLocationDef },
  { folder: 'red-needle', location: RedNeedleLocation as unknown as AuthoredLocationDef },
  { folder: 'rent-a-wreck', location: RentAWreckLocation as unknown as AuthoredLocationDef },
  { folder: 'ruin', location: RuinLocation as unknown as AuthoredLocationDef },
  { folder: 'scrap-yard', location: ScrapYardLocation as unknown as AuthoredLocationDef },
  { folder: 'signal-pit', location: SignalPitLocation as unknown as AuthoredLocationDef },
  { folder: 'skyrail', location: SkyrailLocation as unknown as AuthoredLocationDef },
  { folder: 'supercharging-station', location: SuperchargingStationLocation as unknown as AuthoredLocationDef },
  { folder: 'the-cage', location: TheCageLocation as unknown as AuthoredLocationDef },
  { folder: 'the-meat-market', location: TheMeatMarketLocation as unknown as AuthoredLocationDef },
  { folder: 'the-pineapple-club', location: ThePineappleClubLocation as unknown as AuthoredLocationDef },
];